requirements.txt
torch
transformers
pytesseract
opencv-python
Pillow
langchain
openai
pdf2image
python-dotenv
fastapi
uvicorn
